Add a `name` field to `SourcePath` objects.
Summary:
This diff adds a `name` field to source paths to hold the logical
names that were used to refer to them (e.g. the names that appeared
in BUCK files).  These are set in the `SourcePathTypeCoercer` class
which, for regular paths use the string that was passed into the
coercer, and for build targets, it uses the output "name" given
by the build rule it refers to.  Currently, this only works for
Genrule's which are given a `getOutputName` method to access their
output file names.

Test Plan: included test