Compile tests using ecj.
Summary:
Fix the issues identified by ecj so that we can use ecj to compile our tests.

This generalizes the `<ecj>` macro in `build.xml` so that it can be used in
`<target name="compile-tests">`.

Note that I had to make several changes to `TypeCoercerTest` so that generics
were more explicit to get a test to pass that previously worked with javac but
started failing when being compiled by ecj. Here is an example:

javac
```
assertThat(
    traversal.getObjects(),
    Matchers.<Object>contains(ImmutableList.of(
        sameInstance((Object) input),
        sameInstance((Object) input.get(0)))));
```

ecj
```
assertThat(
    traversal.getObjects(),
    Matchers.contains(ImmutableList.<Matcher<? super Object>>of(
        sameInstance((Object) input),
        sameInstance((Object) input.get(0)))));
```

Note that `Matchers.contains()` is heavily overloaded, as all of the following methods return `Matcher<Iterable<? extends E>>`:

* `Matcher.contains(E...)`
* `Matcher.contains(List<Matcher<? super E>>)`
* `Matcher.contains(Matcher<? super E>...)`
* `Matcher.contains(Matcher<? super E>)`

From the error messages, it appears that javac assumed the second definition but ecj assumed the first definition, which was able to satisfy the second definition, though was less specific.

Test Plan: rm -rf build/testclasses/ && ant compile-tests