Move getBuildFile() to Repository.
Summary:
It's not possible to get the build file for a target unless you have the
repository that that target belongs to. The `getBuildFile()` method on
`BuildTarget` is doing the wrong thing, and all of its callers are broken in
multi-repo situations. Move that method onto `Repository` and update its
callers.

When I'm honest with myself, I think this method should really live on
`RepositoryFactory` (which maybe should be renamed to "something something
Universe"), because only that class has access to the set of *all* repositories
that a target might conceivable come from. However, that does have an icky,
too-much-responsibility feel to it, and right now all the callers have access
to a `Repository`, so this is easier.

Test Plan: buck test --all