Ensure WatchmanWatcher formats query as legal JSON.
Summary:
I'm planning on adding more query parameters to
WatchmanWatcher's query. When I went to do so, I found the query
was constructed by pasting together strings.

Unfortunately, some of the strings (like the root path) come from
outside, so they need to be escaped so the query is valid JSON.

This diff updates WatchmanWatcher to use ObjectMapper to format the
query as legal JSON and adds a test to confirm a previously-invalid
path works.

Test Plan:
Unit test included. Ran buck build buck, confirmed it could talk to Watchman.