Isolate annoation processors used by the in-memory javac to their own classloader.
Summary:
We backed out the update of guava to version 17 since we found that libraries
compiled with an annotation processor that required the processorpath to be set
"leaked" Buck's guava on to the command line. Digging into Oracle's JavaCompiler
class, it looks like it has _something_ to do with the fact that we're letting
javac find and load the annotation processors itself. Fortunately, we can side-
step this by loading themselves from a classloader constructed from the options
given to javac.

Test Plan:
* Update guava to version 17 in Buck, but not in a project built using buck
* Attempt to compile without applying this diff. Build fails.
* Apply this diff
* Attempt to compile with this diff. Build passes.