Improve Javadoc generation to link to Android docs and fail on errors.
Summary:
We recently introduced new Java code that references Android classes in its Javadoc.
It turns out that linking to Android docs is non-trivial because d.android.com does not
include the standard `package-list` file. Instead, Javadoc's `-offline` option must be
used so that things are linked correctly.

Because Buck developers are not required to have the Android SDK installed in order
to develop, this diff changes `build.xml` so there are three Ant-related Javadoc targets:

* `javadoc` Generate Javadoc for Buck source that does not link to the Android docs.
* `javadoc-with-android` Generate Javadoc for Buck source that does links to the Android docs. Requires that `$ANDROID_HOME` is set and points to a local Android SDK. Also, SDK 16 must be installed.
* `verify-javadoc` Similar to `javadoc-with-android`, except `javadoc` is run with the `-private` option and it fails if there are any Javadoc warnings other than the "Error fetching URL" ones that we expect from a system that is offline (such as a CI system in a DMZ).

Because these three targets share a lot of logic in common, new Ant macros were
introduced to avoid copy/paste (`buck-javadoc` and `javadoc-with-android`).

As `verify-javadoc` uncovered a number of Javadoc errors, they have been fixed as part of this diff.

The script to publish Javadocs to the public web site have been updated to use `javadoc-with-android` instead of `javadoc`.

As documented on http://stackoverflow.com/questions/8392559/javadoc-treat-warnings-as-errors,
flagging Javadoc errors is tricky business. Since we have some warnings that we want to whitelist,
this logic has been moved into the new `scripts/verify-javadoc.py` script.

This diff updates the Travis-CI build to run `verify-javadoc` in addition to its previous responsibilities.

Once this diff is submitted, I will update our Sandcastle builds to run
`ant verify-javadoc` instead of `ant javadoc`.

Test Plan:
Introduced multiple Javadoc errors in package-private or private methods
and verified that `ant verify-javadoc` fails when they are present and succeeds
when they are removed.