[iOS] Add optional parameters header_path_prefix and use_buck_header_maps
Summary:
The first parameter is meant to specify a custom prefix for subsequent
library-style imports `#import <PREFIX/foo.h>`.  The default value is
the name of the BUCK target, but some projects use different targets to
generate "flavors" of a same library. The equivalent information is
currently in pieces of xcconfig embedded in BUCK files.

In a next diff I use this new field to set the Xcode build param
PUBLIC_HEADERS_FOLDER_PATH to ../Headers/PREFIX. Another diff will use
the second parameter to generate header maps with Buck.

Test Plan: buildbot