Add AppleExtension rule
Summary:
This adds the "apple extension" rule. These are bundles that get
`.appex` extensions instead of `.app` extensions and get embedded in
other bundles.

In the future this can get refactored into a generic "bundle" rule that
takes the app extension as a parameter, but because I move at a snail's
pace in the Buck codebase, in the interest of unblocking iOS 8 Share
Extension work, and a little bit of YAGNI I'm committing this as-is
(@coneko's original design).