Don't filter out XcodeNative rules from the generated scheme.
Summary:
When fixing up scheme generation in D1494778 I used
isXcodeTargetBuildRuleType to filter out rules that couldn't be
represented as Xcode targets. Unfortunately that does not take
XcodeNative into account, as an XcodeTarget cannot be generated for
those, even though they can be referenced in schemes.

This adds that extra check.

Test Plan: CI, generate a workspace for a target that has XcodeNative
dependencies and build it.