Fix ant tests.
Summary:
When running tests, Buck sets the "user.dir" property to a value of
`Path.toRealPath()` but ant does not. Nor does IJ. On Linux, this
shouldn't be a problem, but on OS X this causes the
`RepositoryFactoryTest` to fail in ant but pass when run with Buck.

Hilarious.

Test Plan:
ant test
buck test --all