Ensure resources used during compilation of java code are released.
Summary:
Turns out that we were opening a `ZipFile` but never calling `close()`
on it when compiling java code in memory from a source zip. This was
causing references to the `ZipEntry` instances to be kept in perm gen
space, ultimately causing it to be exhausted and making Buck fall over,
complaining wildly about that.

Test Plan:
for i in 1 2 3 4 5 6 7 8 9 10; do rm -rf buck-out; buck build buck; done
Attach jvisualvm, look at permgen space. See that's it's not totally
crazy after forcing a gc.
Look at a heapdump and look for ZipEntry instances. There should be none.