Move Parser state into a class of its own.
Summary:
Ongoing work to tidy up the `Parser` a little. Currently, a parser instance is both a
long-lived entity that hangs around for the duration of the buckd process and short-
lived entity that exists to service a single command. By moving the state into its own
class, we start to delineate between these responsibilities.

This is the smallest change I can make --- a follow up diff will move the state to a
top-level class of its own.

Test Plan: buck test --all