Provide a way to explicitly set the `BuildRuleSourcePath` path
Summary:
This diff adds a new constructor to ##BuildRuleSourcePath## so that
the resolved path can be explicitly set, rather than using the output
path of the build rule that is wrapped.  This is useful for build
rules which generate multiple output sources (e.g. thrift, lex, yacc).

Test Plan:
- included test
- Used with C++ lex, yacc, and thrift support, as each of these produce
combinations of C++ sources and headers.