Pass Python launcher nanos epoch to buckd and use when logging events
Summary:
I'm planning on extending the Chrome tracing subsystem out to
the Python launcher, so we can include timing for things like launching
watchmand and buckd itself.

To do this, we need two things:

1) A monotonically-increasing timestamp to write events from Python, and
2) A consistent timestamp epoch between Python and Java

This diff gives us 2) by passing in a new epoch for nanos from Python
to Java when launching `buckd`. When present, we use a new clock
(`com.facebook.buck.timing.NanosAdjustedClock`) in `buckd` which uses the
provided epoch for nanos instead of the default `System.nanoTime()` epoch.

We can get away with this because `System.nanoTime()`'s epoch is undefined,
so we're free to define our own as long as we're consistent about it.

Test Plan:
Edited .bucklogging.local.properties and added
com.facebook.buck.cli.level = FINEST. Ran 'buck project' in buck repo.
Confirmed in logs that new epoch was used. Opened Chrome trace generated
by 'buck project' run in chrome://tracing , confirmed it worked.