Make 'src_target' in xcode_workspace_config optional
Summary:
We'd like to support Xcode workspaces with multiple
test actions and no actual launch / profile actions.

This diff makes the 'src_target' in xcode_workspace_config optional.

However, we previously depended on that for the name of the workspace, so
I added an optional 'workspace_name' parameter, which can be set whether
or not 'src_target' is present.

If it's present, we use it as the name of the workspace, and if not,
we fall back to the short name of 'src_target'. (If that's also not present,
we raise a HumanReadableException).

Test Plan: ant test -Dtest.class=ProjectIntegrationTest