Get the Parser working with cross-repo targets.
Summary:
With these changes, my toy cross-repo example actually makes it
through the Parser (and merrily fails during build step execution).

Very few changes in Parser were actually necessary Just To Get It Working. The
majority of this diff is plumbing a BuildTargetParser down to all the TypeCoercers.
We decided it was better to include this in the constructor than to change the
signature of `coerce` everywhere.

Note that right now we're still using one ProjectBuildFileParser to parse all
the repos. That keeps the changes in Parser smaller for now.

Test Plan: New integration test, `CrossRepoTargetsIntegrationTest`.