Clean up CrossRepoTargetsIntegrationTest to use FakeRepositoryFactory.
Summary:
I wasted a lot of time tracking down a bug that resulted from passing an
empty environment into a RepositoryFactory in this test.
FakeRepositoryFactory should have handled this by default. (Also we need
to get rid of jython, so that we get an outright error if there's no
PATH, but that's a separate diff...)

Test Plan: This test still passes.