Add a rule for downloading files.
Summary:
The URIs for the remote file can either be http, https, or "mvn", which is a
essentially the Buildr maven URL syntax with an optional host added, prefixed
with "mvn:".

By default, Buck disallows builds from downloading artifacts at build time. The
next step in this process will be to create a "fetch" command which will walk
the Target Graph and download all the files that it detects as being necessary.

Test Plan: buck test --all