HttpArtifactCache: add checksum support
Summary:
This diff adds checksum to the HttpArtifactCache, by using the MD5
checksum of the artifact in the "Content-MD5" HTTP header.  On store,
this HTTP header is populated by the client with the checksum formed
from reading the artifact.  On fetch, the header is extracted and
used to compare against the MD5 found after reading the file.

Test Plan:
- included tests
- ran locally with a modified cache server