Change BUCK file python parser to model includes as modules.
Summary:
Previously, include_defs worked by running execfile with the current
context's globals dict to pull the included files globals in.
However, doing includes this way caused some weird behavior:
1) If a build file included two different def files, one after another,
the globals from the first one would be accessible from the context of
the second.
2) Since all include files shared a globals dict with the build file,
they could all overwrite each other's variables, meaning the behavior
of the import was dependent on the import order.
3) Due to how we tried to prepare default gloabls dict for the default
includes, the function closures in the default defs we parsed captured
the original the globals dict.  That meant, that even though we tried
to inject a new globals dict into the includes via execfile, the
functions themselves still maintained references to the original globals
dict (issue #182 on github).  This caused broken behavior when the
functions declared in imports did imports themselves.

This diff changes our include behavior to use modules instead.  When
an include_def is seen, we compile the code in the def into a new
module, using it's own globals dict (seeded with default functions,
etc), then pull the imported include's globals into the callers globals.
As part of this change, we also cache the executed module for each
include, so we only ever execute it once (before, for non-default
includes, we'd execfile the same include everytime we used it).

Fixes #182

Test Plan:
Included some tests for basic sanity, one of which was a test for
https://github.com/facebook/buck/issues/182.