Change ProjectCommand to return ActionGraph instead of PartialGraph from private helper methods.
Summary:
All the methods in ProjectCommand that ultimately call createPartialGraphs don't really
need to use PartialGraph, as they only use ActionGraph. The createPartialGraphs method returns
PartialGraph because it calls out to PartialGraph's static methods to create the ActionGraph
ProjectCommand needs. This changes ProjectCommand so that references to PartialGraph are isolated
in the createPartialGraphs method.

Test Plan: sandcastle