Move PartialGraph#createPartialGraphs method to ProjectCommand.
Summary:
ProjectCommand is the only caller of PartialGraph#createPartialGraphs
that actually needs it. The other callers only call through it to other methods
in PartialGraph by passing parameters to that it does nothing. This moves that
method into ProjectCommand and changes the other callers to call
createPartialGraph directly.

Test Plan: sandcastle