RuleKey: add a `setInput` overload for `SourcePath` objects.
Summary:
This diff adds an overload of ##setInput## for ##SourcePath## objects
that "does the right thing": for ##BuildRuleSourcePath## objects it
hashes the rule key; for ##PathSourcePath## objects it hashes the
contents of the resolved file.

There have been a few rules types where it's important to hash some
input name along with the contents of the input (e.g. C/C++ preprocess
and thrift rules, where the we want to match the location where the
input gets included from along with the contents, in order to correctly
rebuild in the case of renames), so this provides a way to do this
without worrying about what is behind the ##SourcePath##.

Test Plan: included tests