Tracing: delete dangling symlinks before creating new ones.
Summary:
6ab95780 fixed the tracing.py script to work for windows, but also
switched to recreating symlink by first using ##os.path.exists()## to
check if an old version of the symlink needs to be cleaned up before
creating the new one.  However, ##os.path.exists()## uses the ##stat##
syscall under the covers which follows symlinks, and returns ##False##
if the symlink is dangling.

This diff switches to use ##os.path.lexists()##, which uses ##lstat##,
to avoid following the symbolic link, so that we'll still cleanup
dangling symlinks.

Test Plan: arc diff