Fix Xcode project generation always generating all projects and workspaces.
Summary:
The code introduced in D1494778 for searching the roots of the project
generation's target graph when the roots are not specified looks for all
workspace and project rules regardless of whether project or workspace generation
is being run. This fixes that to only search for the correct type of rule.

Test Plan: CI