pathlib: Fix match() API and add more tests
Summary:
When replacing our glob() implementation with pathlib, I ran
into a nasty bug. The match() API (which applies a glob to a path and
returns True or False) didn't work correctly when we passed in globs
like:

  a/**/*.java

I added tests to reproduce the bug and fixed it. I plan to contribute
these fixes upstream as well.

Depends On D1567730

Test Plan: ant py-test