buck_repo.py: Don't suppress subcommand output on error.
Summary:
In several cases in our buck python wrappers, we'd mute the stdout
and stderr subcommands to presumably avoid noise in the common case.
However, this makes it difficult to diagnose issues/errors in those
subcommands (especially via automated build logs).

This diff removes the use of `DEV_NULL` and captures the output
using `subprocess.check_output`, which we print only in the error
case.

Test Plan: Ran locally.