Update tests to use BuildRuleResolver correctly.
Summary:
Because ConstructorArg references BuildRule, Description implementations
can create BuildRule without having to use the provided resolver most of the time.
This means in tests a proper resolver doesn't even need to be passed, a new one can
be created for each Description and everything will work.
Since we're about to change ConstructorArg to reference BuildTarget instead of
BuildRule, this changes the tests to create resolvers and pass them to descriptions
correctly.

Test Plan: CI