JavaTest: don't crash when result file is missing.
Summary:
If a test exits early, it's possible that it doesn't create a test
results file.  Therefore, when we attempt to parse it, we kill buck
with an ugly exception trace.  This diff generates a dummy test case
to signal failure in this case.

Test Plan: Added an integration to handle this case.