Switch back to transitive resource dependencies for Robolectric tests.
Summary:
Apparently, first order deps are enough to get the rules built, but not enough
to actually run them. So switch back to transitive deps, while still using
mini-aapt.

Robolectric also expects the styleable array to be of the same size as the
number of <attr> nodes inside the <declare-styleable> node. It doesn't
matter what the resource ids are though, so just use `0x7f000000`.

Test Plan:
Add a unit test.
sandcastle