More cygwin jiggery-pokery.
Summary:
As suggsted by David Ostrovsky. This introduces the following:

1. git tree can be retrieved for some reasons: fake it for now.
2. path and separator translation between Cygwin and Windows path
   to make Java happy.

I am reliably informed that this gets Buck up and running on Windows, but
not to the point where compilation succeeds, because Python can't be found.
For that to work, the "[tools] -> python" section in `.buckconfig` needs to
be set to (probably) a Windows path.

Test Plan:
Complete hearsay about the ability to run on Cygwin
Still works on OS X.