Fix workspace generation including all projects in all generated workspaces.
Summary:
Workspace used to include projects in the same BUCK files as the targets they
depended on. This meant it would miss projects defined in separate BUCK files.
When this was fixed initially a regression was introduced where workspaces
would include all projects when being generated together. This fixes that and
add a new regression test.

Test Plan: CI, new unit test, run buck project --workspace-and-projects in repo