Treat assumptions as errors
Summary:
We already have an option `test.assumptions-are-errors` that treats an assumption-violated-exception as an error in that it prints the full stack trace.

This diff extends the notion of `assumptions-are-errors` by exiting with a non-zero exit status if assumption-violatied-exceptions occur.

Test Plan:
Updated the test project in `test/com/facebook/buck/junit/testdata/assumption_violations/` with a `.buckconfig` that turns on `tests.assumptions-are-errors`, and checked that the exit status was 42.

If we land this diff I'll write a real test; I'm trying to rush this diff out to help fix a problem in CT-scan.