Allow `BuildRule` objects to opt out of caching
Summary:
There are some class of ##BuildRule## objects which it doesn't make
sense to cache.  The main example right now is the ##SymlinkTree##
rule, which is purely a collection of FS operations which would
happen even if we used caching, and so caching will only ever add
overhead.

This diff adds a ##BuildRule#getCacheMode## method which rules can
override to opt out of caching and changes ##SymlinkTree## to use
this.

Test Plan: added unittest