Remove weak framework support.
Summary:
Currently weak framework support is broken: if a library declares a
strong link on a framework, but the binary using it declares a weak
link, the strong link will override the weak link. Xcode will figure
out whether a framework should be linked to weakly or strongly by
looking at the availability macros. There is no need for a user to
be able to override Xcode's decision, short of frameworks being
annotated with the wrong availability macros. In those rare cases
specifying the linker flags is an acceptable solution.

Test Plan: CI