ThrowableConsoleEvent: include throwable class name in message
Summary:
There have been a few times where the error message alone didn't
provide a good description of the error being reported.  An example
of this is an ##AccessDeniedException##, which only includes the
name of the file being accessed as the message, making it impossible
to figure out the issue without the exception class name.

This diff includes the class name in the error message.

Test Plan: included test