Make SourcePath#asReference protected instead of public.
Summary:
SourcePath#asReference is being used for two unrelated reasons: one
is to have AbstractSourcePath implement some default methods to
reduce boilerplate, the other is to access the value embedded in
the SourcePath, Path in one case and BuildRule in the other. This
changes the method to have protected visibility for the first use
case, and adds a getRelativePath method to PathSourcePath to mirror
BuildRuleSourcePath#getRule for the second one.

Test Plan: CI