Change BuildRuleSourcePath to BuildTargetSourcePath.
Summary:
This changes BuildRuleSourcePath to BuildTargetSourcePath. The class
still functions as a reference to the output of a rule, but it does
not hold a direct reference to the BuildRule, only to a BuildTarget.
This allows it to be created without having to create the BuildRule.
This is needed to be able to move the Description DTO creation to
parsing time, before the TargetGraph is transformed into an
ActionGraph.

Test Plan: CI