Refactor autopopulating description arg fields from createBuildRuleWithDefaults
Summary:
The most verbose parts of creating build rules in tests is populating
Description Arg. This diff refactors the auto-populating code from
createBuildWithDefaults, so that an Arg with defaults can be constructed, and
then overridden. This makes tests more resilient to changes in the Arg, and is
simpler than the old method of passing a modification
function into `createBuildRuleWithDefaults`.

I've applied this to a bunch of apple library descriptions in the tests.

Test Plan: unit tests still pass