Remove uses of RuleJsonPredicate outside Parser.
Summary:
Parser should be the only class dealing with the raw deserialised
JSON maps, all the other parts of the code should deal with higher
level abstractions. This changes all places that use
RuleJsonPredicate to use Predicate<TargetNode<?>> instead. That
does not give access to a higher level abstraction at the moment,
but a subsequent diff will change TargetNode to have a reference
to the Description's DTO.

Test Plan: CI