Do not cache rules which have input files under symlinks
Summary:
Our `Parser` cache currently assumes that all input files
for a BUCK file exist in the same directory as the BUCK file or
a subdirectory.

Because of this, if a rule depends on input files (say,
`glob(['foo/*.java'])` and `foo` is actually a symbolic link to
somewhere else in the repository (say `bar`), the `Parser` cache has
no idea that Watchman events for paths like `bar/x/y/Z.java` have to
invalidate rules that depend on `foo/A.java`.

Since we do not pass the `glob` pattern up from Python to Java, it's
impossible for us to determine in Java that a Watchman event for
`bar/x/y/Z.java` is related to `foo/*.java`.

So, this diff adds detection that `foo/A.java` is in a path with a
component that is a symbolic link, warns if it detects this scenario,
then blows away the cached rule(s) referring to this path after each
command.

Test Plan:
Integration tests included. Tests all failed before my diff
(commented out explicit call to `Parser.cleanCache()`), now they pass.