HttpArtifactCache: don't set "filename" in content disposition
Summary:
The jetty server code has an apparent bug where, if this is set,
it *always* creates local files to store the contents for multi-
part requests, regardless of the config setting we pass in.  So,
don't set this on the client.

Test Plan:
Ran a build against a HTTP cache server and verified /tmp/MultiPart*
files were not created on cache sets.