Prompt user to kill Xcode when running 'buck project'
Summary:
Users are getting confused when `buck project` writes
to files which Xcode has open. Xcode usually starts spinning
and crashing when this happens.

This diff builds on D1628403 to prompt the user to quit Xcode
if it's running when running `buck project`.

Test Plan:
`buck project` in Xcode repo with Xcode running and without.
With Xcode running, I was prompted to quit it. Without, I was not prompted.

Unfortunately, there are no unit or integration tests for
ProjectCommand.