Shuffle tests before passing to the test runner
Summary:
Add a `--shuffle` flag to `buck test` which, surprisingly, shuffles the test class order (only for Java tests).

If `--shuffle` is not passed, the order is determined by a hardcoded seed. Always reordering the test classes (even with a hardcoded seed) decouples the testing order from the filesystem and environment.

Test Plan:
I have manually verified the following facts:
 1. The order without `--shuffle` is not the same as the order before this diff
 2. The order without `--shuffle` is consistent.
 3. The order with `--shuffle` changes every run.
 4. `--shuffle` works with test selectors (no reason it wouldn't, selection is part of the test runner)