Add an option for `[apple] xcode_developer_dir` in .buckconfig.
Summary:
This will be used in an upcoming integration test for the `#compilation-database` flavor.
In practice, Buck's default of using `xcode-select --print-path` will work fine for most
people.

Test Plan:
I have a pending diff that builds the `#compilation-database` flavor.
In my iOS directory, I added a `.buckconfig.local` with:

```
[apple]
  xcode_developer_dir = /totally/fakeplace
```

I rebuilt my compilation database and verified that paths were prefixed
with `/totally/fakeplace` in the output, as appropriate.