Amend the StubJar to also handle directories of class files.
Summary:
There's an assumption that compiling code using the `JavacStep` also generates
the ABI of the compiled code. In order to support the pattern of calculating
the stub jar from the compiled output before jarring, change the `StubJar` to
seamlessly handle the case where the given `Path` is a directory.

Test Plan: buck test --all