buck targets prints nothing if every --referenced_file is outside the project.
Summary:
Previously, if a `--referenced_file` were outside the project root, it would
be ignored. If there were only one `--referenced_file`, then this would behave
the same as if no `--referenced_file` arguments were passed, which meant that
all targets in the project would be returned, which is counterintuitive.

Test Plan: Introduced `TargetsCommandIntegrationTest` and updated related tests.