Fix malformed Javadoc.
Summary:
Apparently our current Javadoc checks do not catch this because we only
run it on the src/ folder since we only generate public Javadocs for src/.
I am working on extending this to the test/ folder in a separate diff.

Test Plan: Yellow warning icons go away in Eclipse.