Escape strings in argfiles passed to external javac
Summary:
The arguments written to javac's @argfiles need to be escaped if
they contain some special characters.  The `#` character is particularly
important to escape as we use this for flavors and, therefore, will show
in generated path names quite frequently.

This diff generalizes the escaper logic a bit and creates and escaper
for @argfiles.

Test Plan: added a test