Add buckconfig option to disable Xcode process check
Summary:
Advanced engineers asked for the ability to disable Buck's
prompt to kill the running IDE. They said they prefer dealing with
IDE instability to having to kill and relaunch the (extremely slow)
IDE.

This diff adds a new setting to `.buckconfig` / `.buckconfig.local`:

  [project]
    ide_prompt = false

to let advanced users disable the check. We let users know about
this setting after they answer the prompt.

I also tidied up the logic in `ProjectCommand` to make it more readable
and fix a bug where we incorrectly logged that we couldn't check if
Xcode was running in the case where Xcode was not running.

Test Plan:
`arc build` in Xcode repo with Xcode open and with it closed,
with and without new setting in `.buckconfig.local`.