Fix ProcessExecutor so it doesn't print empty strings on process failure
Summary:
I noticed an earlier diff caused SuperConsole to stop printing output,
even when nothing was printed to the console (which I added in a followup
diff).

I added debug logging and found a bug: ProcessExecutor calls
`OutputStream.print("")` if the process does not exit cleanly, even if
the process prints nothing to stdout or stderr.

Since calling `print()` on `DirtyPrintStreamDecorator` has the side
effect of causing SuperConsole to cease updating, this is pretty bad.
(It means any time we pass a process to `ProcessExecutor` and it fails
silently, we stop the console completely.)

This diff fixes the bug and adds a test which reproduces the bug and confirms
the fix.

Test Plan:
Unit test included.
Ran `ant java-test -Dtest.class=ProcessExecutorTest` before fix, confirmed
it failed. Applied fix, repeated test, confirmed it passed.