Support multi-device install for exopackages
Summary:
Most functionality (including all non-final fields) has been moved into
a private inner class that is instantiated per-device.  An AtomicInteger
is used to allocate a separate port number per-device.  Static methods
were not moved to allow for testing, and because inner classes cannot
have static methods.

Test Plan:
Ran a -x install on 2 devices and 2 emulators (all at once).  All were
updated properly.  All installations showed up in the trace and the log,
though without identifiers for which was which.  Initially, I forgot to
move nativeAgentPath to the inner class, and I got a failure because it
was different on the different devices.  The exception properly showed
the id of the failed device.