New utility class AppleSdkDiscovery
Summary:
To build the `AppleCxxPlatform`, we need to discover the
location of the components of the Apple SDK.

This introduces a new utility class `AppleSdkDiscovery` which takes
the path to the Xcode installation (which I will get from @mbolin's
D1630292) and discovers the Mac OS X, iPhone OS, and iPhone Simulator
SDKs inside.

It also reads the `SDKSettings.plist` file inside each SDK so we can
refer to them by the same name as shown by `xcodebuild -showsdks`:

  % xcodebuild -showsdks
  OS X SDKs:
     OS X 10.9                            -sdk macosx10.9

  iOS SDKs:
      iOS 8.0                             -sdk iphoneos8.0

  iOS Simulator SDKs:
      Simulator - iOS 8.0                 -sdk iphonesimulator8.0

Test Plan: Unit test included.