Easy: Fix ant build when immutables used in same library they're defined in
Summary:
I hit a bug using a code-generated immutable in the same
library it was defined in. This was because we weren't compiling the
code-generated immutables before using them.

I decided to follow what we do for javacc and compile them with javac
before the rest of the ecj run. This fixed the issue with the ant build.

Test Plan:
Changed //src/com/facebook/buck/apple:rules to java_immutables_library()
and added an immutable type which I used in another class in that package. Before,
`ant default` failed. Now, it works.