Move a path args utility from TargetsCommandOptions to PathArguments.
Summary:
Parsing arguments of type `Path` is a general issue that we need to solve in Buck.
args4j has `org.kohsuke.args4j.spi.PathOptionHandler`, but it does not do enough
to resolve `Path` arguments to absolute paths or anything like that.

This diff is a pure refactoring: a subsequent diff will use it for `buck audit owner`.

Test Plan: Sandcastle