Alternative minimal tempfile implementation
Summary:
We make fairly minimal usage of the actual tempfile implementation. Due
to the differences between Win32 and Unix implementations of the module
it was necessary to create a minimal implementation that meets our
requirements to allow our use of it to work on Windows.

Test Plan: Run buck on Windows and Unix