Change SeparatedProjectsGenerator to get a TargetGraph instead of an ActionGraph.
Summary:
Project generation shouldn't need to construct the action graph. This
changes SeparatedProjectsGenerator to take a TargetGraph parameter
instead of an ActionGraph parameter in its constructor. It does not
change anything else, but it sets up the stage for changing it to not
need a to create an ActionGraph.

Test Plan: CI