Move testShouldNotPolluteClasspathWhenProcessorPathIsSet to Python
Summary:
Currently this test is written in Java, and it launches Buck with
ProjectWorkspace.runBuckBuild, which works directly with the Main
class. This is fine today, because the current fix to the bug is
entirely on the Java side.

We're about to change how the fix works, and part of it will be
in the Python launcher. The test code will thus have to launch Buck
that way rather than working with the Main class directly.
This diff rewrites the test to invoke the Buck launcher.

Test Plan:
Run the new test, verify passes
Introduce a failure, verify fails